added BG subset sampling for DS generation

+++ src/create_yolo_ds.py
@@ -11,6 +11,8 @@ import matplotlib.pyplot as plt
 from tqdm.auto import tqdm
 from yolo_utils import plot_yolo_sample
 
+random.seed(2022)
+
 def exists_or_make_dir(d):
     if not os.path.exists(d):
         os.mkdir(d)
@@ -26,7 +28,11 @@ exists_or_make_dir(IMG_DIR)
 exists_or_make_dir(ANNO_DIR)
 df = pd.read_csv('./input/train_folds.csv')
 
-fold_fns = [[] for _ in range(5)]
+fold_all_fns = [[] for _ in range(5)]
+fold_fg_fns = [[] for _ in range(5)]
+fold_bg_fns = [[] for _ in range(5)]
+
+BG_KEEP_EVERY_N = 40
 
 for idx in tqdm(range(len(df))):
     video_id = df.iloc[idx]['video_id']
@@ -42,30 +48,36 @@ for idx in tqdm(range(len(df))):
     image_id = df.iloc[idx]['image_id']
     new_fn = os.path.join(IMG_DIR, image_id + ".jpg")
 
-    fold_fns[fold].append(new_fn)
-
     # create a symlink for image
-    os.symlink(os.path.abspath(img_fn), new_fn)
+    if not os.path.exists(new_fn):
+        os.symlink(os.path.abspath(img_fn), new_fn)
 
     # create yolo annotations
     label_fn = os.path.join(ANNO_DIR, image_id + '.txt')
     annos = eval(df.iloc[idx]['annotations'])
+
+    fold_all_fns[fold].append(new_fn)
+    if len(annos) > 0:
+        fold_fg_fns[fold].append(new_fn)
+    else:
+        fold_bg_fns[fold].append(new_fn)
     
-    with open(label_fn, 'w') as f:
-        for anno in annos:
-            x,y = int(anno['x']), int(anno['y'])
-            w,h = int(anno['width']), int(anno['height'])
+    if not os.path.exists(label_fn):
+        with open(label_fn, 'w') as f:
+            for anno in annos:
+                x,y = int(anno['x']), int(anno['y'])
+                w,h = int(anno['width']), int(anno['height'])
 
-            # scale to 0-1 coordinates
-            x,y,w,h = (
-                np.array([x,y,w,h]).astype(np.float) /
-                np.array([im_w, im_h, im_w, im_h]).astype(np.float)
-                )
-            xc = x + w / 2
-            yc = y + h / 2
+                # scale to 0-1 coordinates
+                x,y,w,h = (
+                    np.array([x,y,w,h]).astype(np.float) /
+                    np.array([im_w, im_h, im_w, im_h]).astype(np.float)
+                    )
+                xc = x + w / 2
+                yc = y + h / 2
 
-            line = f'0 {xc:.4f} {yc:.4f} {w:.4f} {h:.4f}'
-            f.write(line + '\n')
+                line = f'0 {xc:.4f} {yc:.4f} {w:.4f} {h:.4f}'
+                f.write(line + '\n')
 
 # save training and val file paths
 for fold in range(5):
@@ -74,9 +86,16 @@ for fold in range(5):
     val_fns = []
     for i in range(5):
         if i != fold:
-            train_fns += fold_fns[i].copy()
+            # select all FGs
+            fg_fns = fold_fg_fns[i].copy()
+            # select every nth bg randomly
+            bg_fns = fold_bg_fns[i].copy()
+            random.shuffle(bg_fns)
+
+            train_fns += fg_fns + bg_fns[::BG_KEEP_EVERY_N]
         else:
-            val_fns += fold_fns[i].copy()
+            # keep all samples for validation
+            val_fns += fold_all_fns[i].copy()
     
     # shuffle training file order
     random.shuffle(train_fns)
@@ -110,5 +129,5 @@ for fold in range(5):
             default_flow_style=False)
 
 # Check that conversion worked
-plot_yolo_sample(fold_fns[0][10])
+plot_yolo_sample(fold_fg_fns[0][0])
 plt.show()
